+++ app/main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"app/database"
 	"app/models"
-	"strconv"
 
 	"github.com/gin-gonic/gin"
 )
@@ -23,23 +22,6 @@ func main() {
 	r.GET("/book/:id/Content", apiGetContent)
 	r.Run()
 }
-func apiGetBook(c *gin.Context) {
-	println("get all posts")
-	param := c.Param("id")
-	intParam, _ := strconv.Atoi(param)
-	posts := models.GetBook(database.GetDB(), intParam)
-	c.JSON(200, posts)
-}
-func apiGetAllBook(c *gin.Context) {
-	posts := models.GetAllBook(database.GetDB())
-	c.JSON(200, posts)
-}
-func apiGetContent(c *gin.Context) {
-	param := c.Param("id")
-	intParam, _ := strconv.Atoi(param)
-	posts := models.GetBookContent(database.GetDB(), intParam)
-	c.JSON(200, posts)
-}
 func getAllPost(c *gin.Context) {
 	println("get all posts")
 	posts := models.GetAllPosts(database.GetDB())
@@ -50,7 +32,6 @@ func getAllPost(c *gin.Context) {
 
 	c.JSON(200, posts)
 }
-
 func createPost(c *gin.Context) {
 	post := &models.Post{}
 	err := c.Bind(post)
